+++ packages/svelte-form/src/index.ts
@@ -1,6 +1,6 @@
 export * from '@tanstack/form-core'
 
-export { useStore } from '@tanstack/svelte-store'
+export { useSelector, useStore } from '@tanstack/svelte-store'
 
 export { createForm, type SvelteFormApi } from './createForm.svelte.js'